# survey-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey app api",
        "version": "0.1"
    },
    "paths": {
        "/api/test": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Fetch test data",
                "description": "Fetch test data.",
                "operationId": "7c2e6b5602b42636e41cc7228a9d8765",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/test/submit": {
            "post": {
                "tags": [
                    "Test"
                ],
                "summary": "Submit test data",
                "operationId": "46e9edb50f0b85a545a47d0916f82702",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "responses"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "example@test.com"
                                    },
                                    "responses": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "question_id"
                                            ],
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer"
                                                },
                                                "option_id": {
                                                    "type": "integer"
                                                },
                                                "user_input": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}